refactor: split typings.ts into multiple files
