+++ src/requests/typings/appAction.ts
@@ -1,0 +1,11 @@
+import { PlainClientAPI } from 'contentful-management'
+
+export type AppActionCallContext = {
+  cma: PlainClientAPI
+  appActionCallContext: {
+    spaceId: string
+    environmentId: string
+    appInstallationId: string
+    userId: string
+  }
+}
